Add support for thumbnails and deleting photos.
This includes a lot of new path management code, which is hideous.  Next
step is to refactor this into proper, separate solutions for
'Production' (backed by S3) and 'development' (backed by local storage).

As usual, UI is ugly and painful, with nothing to inform the user of
upload progress.  That comes later.