Ugly rough cut at a page to show photos by tag.
This recycles the stream template, for now.  This is mostly to drive
testing & exploration of a client side URL routing solution (necessary
if if I'm to ditch server side templating...).  Maybe keep it simple:
continue to generate full URLs server side, but embed them in each JSON
request?  Too limited?